third_party/blink/web_tests/fast/dom/shadow/suppress-mutation-events-in-shadow-characterdata-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMCharacterDataModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
Ensures that DOMCharacterDataModified isn't fired inside shadows.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fired is false
PASS div.innerHTML is 'Hello, World!'
PASS successfullyParsed is true

TEST COMPLETE
